Refactor: 공통 모듈 위치 이동
