+++ src/common/containers/SolveViewWrapper/SolveViewWrapper.css
